+++ apps/website/src/examples/DropdownButton.main.tsx
@@ -6,18 +6,14 @@ import * as React from 'react';
 import { MenuItem, DropdownButton } from '@itwin/itwinui-react';
 
 export default () => {
-  const onClick = (index: number, close: () => void) => () => {
-    close();
-  };
-
   const buttonMenuItems = (close: () => void) => [
-    <MenuItem key={1} onClick={onClick(1, close)}>
+    <MenuItem key={1} onClick={() => close()}>
       Item #1
     </MenuItem>,
-    <MenuItem key={2} onClick={onClick(2, close)}>
+    <MenuItem key={2} onClick={() => close()}>
       Item #2
     </MenuItem>,
-    <MenuItem key={3} onClick={onClick(3, close)}>
+    <MenuItem key={3} onClick={() => close()}>
       Item #3
     </MenuItem>,
   ];