+++ apps/website/src/examples/index.tsx
@@ -3,56 +3,126 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 export { default as AlertMainExample } from './Alert.main';
+export { default as AlertInlineExample } from './Alert.inline';
+export { default as AlertInformationalExample } from './Alert.informational';
+export { default as AlertPositiveExample } from './Alert.positive';
+export { default as AlertNegativeExample } from './Alert.negative';
+export { default as AlertStickyExample } from './Alert.sticky';
+export { default as AlertWarningExample } from './Alert.warning';
+
 export { default as AvatarMainExample } from './Avatar.main';
+export { default as AvatarGroupExample } from './Avatar.group';
+export { default as AvatarGroupAnimatedExample } from './Avatar.groupanimated';
+export { default as AvatarGroupOverflowExample } from './Avatar.groupoverflow';
+export { default as AvatarGroupStackedExample } from './Avatar.groupstacked';
+export { default as AvatarIconExample } from './Avatar.icon';
+export { default as AvatarInitialsExample } from './Avatar.initials';
+export { default as AvatarPictureExample } from './Avatar.picture';
+export { default as AvatarSizesExample } from './Avatar.sizes';
+export { default as AvatarStatusesExample } from './Avatar.statuses';
+
 export { default as AnchorMainExample } from './Anchor.main';
+
 export { default as BadgeMainExample } from './Badge.main';
+export { default as BadgeSoftExample } from './Badge.soft';
+export { default as BadgeStatusesExample } from './Badge.statuses';
+
 export { default as BlockquoteMainExample } from './Blockquote.main';
-export { default as ButtonMainExample } from './Button.main';
-export { default as ButtonDefaultExample } from './Button.default';
-export { default as IconButtonMainExample } from './IconButton.main';
-export { default as DropdownButtonMainExample } from './DropdownButton.main';
-export { default as ButtonGroupMainExample } from './ButtonGroup.main';
-export { default as SplitButtonMainExample } from './SplitButton.main';
-export { default as IdeasButtonMainExample } from './IdeasButton.main';
+
+export { default as BreadcrumbsButtonExample } from './Breadcrumbs.button';
+export { default as BreadcrumbsExtremeTruncationExample } from './Breadcrumbs.extremeTruncation';
+export { default as BreadcrumbsFolderExample } from './Breadcrumbs.folder';
+export { default as BreadcrumbsLinkExample } from './Breadcrumbs.link';
 export { default as BreadcrumbsMainExample } from './Breadcrumbs.main';
-export { default as CheckboxMainExample } from './Checkbox.main';
+export { default as BreadcrumbsTruncationExample } from './Breadcrumbs.truncation';
+
+export { default as ButtonDefaultExample } from './Button.default';
+export { default as ButtonMainExample } from './Button.main';
+export { default as IconButtonMainExample } from './IconButton.main';
+export { default as IdeasButtonMainExample } from './IdeasButton.main';
+export { default as SplitButtonMainExample } from './SplitButton.main';
+
+export { default as ButtonGroupMainExample } from './ButtonGroup.main';
+
 export { default as CarouselMainExample } from './Carousel.main';
-export { default as ColorPickerMainExample } from './ColorPicker.main';
-export { default as ComboBoxMainExample } from './ComboBox.main';
+
+export { default as CheckboxMainExample } from './Checkbox.main';
+
 export { default as CodeMainExample } from './Code.main';
+
+export { default as ColorPickerMainExample } from './ColorPicker.main';
+
+export { default as ComboBoxMainExample } from './ComboBox.main';
+
 export { default as DatePickerMainExample } from './DatePicker.main';
-export { default as DialogStaticExample } from './Dialog.static';
+
+export { default as DialogDismissibleExample } from './Dialog.dismissible';
+export { default as DialogDraggableExample } from './Dialog.draggable';
+export { default as DialogFullPageExample } from './Dialog.fullpage';
 export { default as DialogMainExample } from './Dialog.main';
+export { default as DialogModalExample } from './Dialog.modal';
+export { default as DialogNonDismissibleExample } from './Dialog.nondismissible';
+
+export { default as DropdownButtonMainExample } from './DropdownButton.main';
+
 export { default as DropdownMenuMainExample } from './DropdownMenu.main';
-export { default as NonIdealStateMainExample } from './NonIdealState.main';
+
 export { default as ExpandableBlockMainExample } from './ExpandableBlock.main';
+
 export { default as FieldsetMainExample } from './Fieldset.main';
+
 export { default as FileUploadMainExample } from './FileUpload.main';
+
+export { default as FlexMainExample } from './Flex.main';
+
 export { default as FooterMainExample } from './Footer.main';
+
 export { default as HeaderMainExample } from './Header.main';
+
 export { default as InformationPanelMainExample } from './InformationPanel.main';
-export { default as KeyboardMainExample } from './Keyboard.main';
+
+export { default as InputGroupMainExample } from './InputGroup.main';
+
 export { default as InputMainExample } from './Input.main';
+
+export { default as KeyboardMainExample } from './Keyboard.main';
+
 export { default as LabelMainExample } from './Label.main';
+
+export { default as NonIdealStateMainExample } from './NonIdealState.main';
+
 export { default as ProgressLinearMainExample } from './ProgressLinear.main';
 export { default as ProgressRadialMainExample } from './ProgressRadial.main';
+
 export { default as RadioMainExample } from './Radio.main';
 export { default as RadioTileMainExample } from './RadioTile.main';
+
 export { default as SelectMainExample } from './Select.main';
-export { default as InputGroupMainExample } from './InputGroup.main';
+
 export { default as SideNavigationMainExample } from './SideNavigation.main';
+
 export { default as SliderMainExample } from './Slider.main';
-export { default as SurfaceMainExample } from './Surface.main';
-export { default as ToastStaticExample } from './Toast.static';
-export { default as ToastMainExample } from './Toast.main';
-export { default as ToggleSwitchMainExample } from './ToggleSwitch.main';
-export { default as TableMainExample } from './Table.main';
-export { default as TabsMainExample } from './Tabs.main';
-export { default as TextMainExample } from './Text.main';
-export { default as TagMainExample } from './Tag.main';
-export { default as TextareaMainExample } from './Textarea.main';
-export { default as TileMainExample } from './Tile.main';
-export { default as TooltipMainExample } from './Tooltip.main';
-export { default as TooltipStaticExample } from './Tooltip.static';
-export { default as TreeMainExample } from './Tree.main';
+
 export { default as StepperMainExample } from './Stepper.main';
+
+export { default as SurfaceMainExample } from './Surface.main';
+
+export { default as TableMainExample } from './Table.main';
+
+export { default as TabsMainExample } from './Tabs.main';
+
+export { default as TagMainExample } from './Tag.main';
+
+export { default as TextareaMainExample } from './Textarea.main';
+
+export { default as TextMainExample } from './Text.main';
+
+export { default as TileMainExample } from './Tile.main';
+
+export { default as ToastMainExample } from './Toast.main';
+
+export { default as ToggleSwitchMainExample } from './ToggleSwitch.main';
+
+export { default as TooltipMainExample } from './Tooltip.main';
+
+export { default as TreeMainExample } from './Tree.main';