+++ packages/itwinui-react/src/core/utils/hooks/index.ts
@@ -14,3 +14,5 @@ export * from './useSafeContext';
 export * from './useLatestRef';
 export * from './useIsomorphicLayoutEffect';
 export * from './useIsThemeAlreadySet';
+export * from './useIsClient';
+export * from './useId';